% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puma_fact.R
\name{puma_fact}
\alias{puma_fact}
\title{Tells you a Puma Fun Fact}
\usage{
puma_fact(puma_name, weight, height)
}
\arguments{
\item{puma_name}{A character containing the input puma name.}

\item{weight}{A number that indicates weight of the puma specified by puma name.}

\item{height}{A number that indicates height of the puma specified by puma name.}
}
\value{

}
\description{
Contains one function about a specific puma based on their name, weight, and height.
}
