% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puma_fact.R
\name{puma_mood}
\alias{puma_mood}
\title{Mood of Pumas During the Week}
\usage{
puma_mood(weekday, mood)
}
\arguments{
\item{weekday}{This is a string that represents the days of the week (i.e. Monday, Tuesday, Wednesday, Thursday, Friday etc).}

\item{mood}{This is a string that represents the mood of the Puma (i.e. happy or sad).}
}
\value{

}
\description{
Mood of Pumas During the Week
}
