% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puma_fact.R
\name{puma_warning}
\alias{puma_warning}
\title{Puma Warning}
\usage{
puma_warning(distance, presence)
}
\arguments{
\item{distance}{A numeric value indicating distance in feet between person and puma.}

\item{presence}{A character value indicating whether or not a puma is present. Explicitly use "YES" or "NO".}
}
\value{

}
\description{
This function is setup to provide hikers with a puma warning based on puma presence and distance (ft).
}
