% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lakers.R
\name{theme_lakers}
\alias{theme_lakers}
\title{theme_lakers}
\usage{
theme_lakers()
}
\value{

}
\description{
This is a custom theme lightly inspired by the NBA Team, Lakers. Colors include: yellow, navy blue, light gray, and purple. Use theme() if you need to tweak this theme.
}
